feat: 注册接口

+++ src/models/resp.rs
@@ -4,7 +4,7 @@ use serde_json::Value;
 #[derive(Deserialize, Serialize, Debug, Clone)]
 pub(crate) struct Response {
     #[serde(rename = "code")]
-    pub code: u32,
+    pub code: u16,
     #[serde(rename = "message")]
     pub message: Option<String>,
     #[serde(rename = "data")]
